Fixes to build and work on sles15 and not break on old releases.

+++ dep/build_ssh.sh
@@ -63,7 +63,8 @@ for dir in $dirs; do
             # SLES has symlinks for asm
             realpath=$(readlink -f "/usr/include/$dir")
             cp -rp "$realpath" "${SPRT_PREFIX}/include/"
-	else
+	fi
+        if [ ! -e "${SPRT_PREFIX}/include/$dir" ]; then
             cp -rp "/usr/include/$dir" "${SPRT_PREFIX}/include/"
         fi
     fi
